pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "call-center-ai"
version = "0.1.0"
description = "AI-powered call center for tire shop (Ukraine)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # LLM
    "anthropic>=0.42.0",
    # Google Cloud
    "google-cloud-speech>=2.28.0",
    "google-cloud-texttospeech>=2.21.0",
    # HTTP client
    "aiohttp>=3.11.0",
    # Database
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    # Redis
    "redis[hiredis]>=5.2.0",
    # Resilience
    "aiobreaker>=1.2.0",
    # Monitoring
    "prometheus-client>=0.21.0",
    # Config
    "pydantic-settings>=2.7.0",
    # Task queue
    "celery[redis]>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "aioresponses>=0.7.7",
]
whisper = [
    "faster-whisper>=1.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = [
    "E501",    # line too long
    "RUF001",  # ambiguous unicode (Cyrillic in Ukrainian text)
    "RUF002",  # ambiguous unicode in docstrings
    "RUF003",  # ambiguous unicode in comments
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "aiobreaker.*",
    "google.cloud.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration (deselect with '-m \"not integration\"')",
    "e2e: marks tests as end-to-end (deselect with '-m \"not e2e\"')",
    "load: marks tests as load tests (deselect with '-m \"not load\"')",
]
